fix(sui-transactional-test-runner): compilation errors were fixed
