+++ crates/sui-transactional-test-runner/Cargo.toml
@@ -35,6 +35,8 @@ move-stdlib = { path = "../../external-crates/move/crates/move-stdlib" }
 move-vm-runtime = { path = "../../external-crates/move/crates/move-vm-runtime" }
 
 rocksdb.workspace = true
+simulacrum.workspace = true
+sui-genesis-builder.workspace = true
 sui-graphql-rpc.workspace = true
 sui-rest-api.workspace = true
 sui-framework.workspace = true
