+++ crates/sui-transactional-test-runner/src/lib.rs
@@ -8,42 +8,23 @@
 
 pub mod args;
 pub mod programmable_transaction_test_parser;
-mod simulator_persisted_store;
 pub mod test_adapter;
 
 pub use move_transactional_test_runner::framework::run_test_impl;
 use rand::rngs::StdRng;
-use simulacrum::Simulacrum;
-use simulacrum::SimulatorStore;
-use simulator_persisted_store::PersistedStore;
+use simulacrum::{InMemoryStore, Simulacrum};
 use std::path::Path;
-use std::sync::Arc;
-use sui_core::authority::authority_test_utils::send_and_confirm_transaction_with_execution_error;
-use sui_core::authority::AuthorityState;
-use sui_json_rpc::authority_state::StateRead;
 use sui_json_rpc_types::DevInspectResults;
-use sui_json_rpc_types::EventFilter;
-use sui_storage::key_value_store::TransactionKeyValueStore;
-use sui_types::base_types::ObjectID;
 use sui_types::base_types::SuiAddress;
-use sui_types::base_types::VersionNumber;
 use sui_types::digests::TransactionDigest;
-use sui_types::digests::TransactionEventsDigest;
 use sui_types::effects::TransactionEffects;
-use sui_types::effects::TransactionEvents;
 use sui_types::error::ExecutionError;
-use sui_types::error::SuiError;
 use sui_types::error::SuiResult;
 use sui_types::event::Event;
-use sui_types::messages_checkpoint::CheckpointContentsDigest;
 use sui_types::messages_checkpoint::VerifiedCheckpoint;
-use sui_types::object::Object;
-use sui_types::storage::ObjectStore;
 use sui_types::storage::ReadStore;
 use sui_types::sui_system_state::epoch_start_sui_system_state::EpochStartSystemStateTrait;
-use sui_types::sui_system_state::SuiSystemStateTrait;
 use sui_types::transaction::Transaction;
-use sui_types::transaction::TransactionDataAPI;
 use sui_types::transaction::TransactionKind;
 use test_adapter::{SuiTestAdapter, PRE_COMPILED};
 
@@ -53,12 +34,6 @@ pub async fn run_test(path: &Path) -> Result<(), Box<dyn std::error::Error>> {
     telemetry_subscribers::init_for_testing();
     run_test_impl::<SuiTestAdapter>(path, Some(&*PRE_COMPILED)).await?;
     Ok(())
-}
-
-pub struct ValidatorWithFullnode {
-    pub validator: Arc<AuthorityState>,
-    pub fullnode: Arc<AuthorityState>,
-    pub kv_store: Arc<TransactionKeyValueStore>,
 }
 
 #[allow(unused_variables)]
@@ -102,251 +77,7 @@ pub trait TransactionalAdapter: Send + Sync + ReadStore {
 }
 
 #[async_trait::async_trait]
-impl TransactionalAdapter for ValidatorWithFullnode {
-    async fn execute_txn(
-        &mut self,
-        transaction: Transaction,
-    ) -> anyhow::Result<(TransactionEffects, Option<ExecutionError>)> {
-        let with_shared = transaction
-            .data()
-            .intent_message()
-            .value
-            .contains_shared_object();
-        let (_, effects, execution_error) = send_and_confirm_transaction_with_execution_error(
-            &self.validator,
-            Some(&self.fullnode),
-            transaction,
-            with_shared,
-        )
-        .await?;
-        Ok((effects.into_data(), execution_error))
-    }
-
-    async fn dev_inspect_transaction_block(
-        &self,
-        sender: SuiAddress,
-        transaction_kind: TransactionKind,
-        gas_price: Option<u64>,
-    ) -> SuiResult<DevInspectResults> {
-        self.fullnode
-            .dev_inspect_transaction_block(
-                sender,
-                transaction_kind,
-                gas_price,
-                None,
-                None,
-                None,
-                None,
-                None,
-            )
-            .await
-    }
-
-    async fn query_tx_events_asc(
-        &self,
-        tx_digest: &TransactionDigest,
-        limit: usize,
-    ) -> SuiResult<Vec<Event>> {
-        Ok(self
-            .validator
-            .query_events(
-                &self.kv_store,
-                EventFilter::Transaction(*tx_digest),
-                None,
-                limit,
-                false,
-            )
-            .await
-            .unwrap_or_default()
-            .into_iter()
-            .map(|sui_event| sui_event.into())
-            .collect())
-    }
-
-    async fn create_checkpoint(&mut self) -> anyhow::Result<VerifiedCheckpoint> {
-        unimplemented!("create_checkpoint not supported")
-    }
-
-    async fn advance_clock(
-        &mut self,
-        _duration: std::time::Duration,
-    ) -> anyhow::Result<TransactionEffects> {
-        unimplemented!("advance_clock not supported")
-    }
-
-    async fn advance_epoch(&mut self, _create_random_state: bool) -> anyhow::Result<()> {
-        unimplemented!("advance_epoch not supported")
-    }
-
-    async fn request_gas(
-        &mut self,
-        _address: SuiAddress,
-        _amount: u64,
-    ) -> anyhow::Result<TransactionEffects> {
-        unimplemented!("request_gas not supported")
-    }
-
-    async fn get_active_validator_addresses(&self) -> SuiResult<Vec<SuiAddress>> {
-        Ok(self
-            .fullnode
-            .get_system_state()
-            .map_err(|e| {
-                SuiError::SuiSystemStateReadError(format!(
-                    "Failed to get system state from fullnode: {}",
-                    e
-                ))
-            })?
-            .into_sui_system_state_summary()
-            .active_validators
-            .iter()
-            .map(|x| x.sui_address)
-            .collect::<Vec<_>>())
-    }
-}
-
-impl ReadStore for ValidatorWithFullnode {
-    fn get_committee(
-        &self,
-        _epoch: sui_types::committee::EpochId,
-    ) -> sui_types::storage::error::Result<Option<Arc<sui_types::committee::Committee>>> {
-        todo!()
-    }
-
-    fn get_latest_checkpoint(&self) -> sui_types::storage::error::Result<VerifiedCheckpoint> {
-        let sequence_number = self
-            .validator
-            .get_latest_checkpoint_sequence_number()
-            .unwrap();
-        self.get_checkpoint_by_sequence_number(sequence_number)
-            .map(|c| c.unwrap())
-    }
-
-    fn get_highest_verified_checkpoint(
-        &self,
-    ) -> sui_types::storage::error::Result<VerifiedCheckpoint> {
-        todo!()
-    }
-
-    fn get_highest_synced_checkpoint(
-        &self,
-    ) -> sui_types::storage::error::Result<VerifiedCheckpoint> {
-        todo!()
-    }
-
-    fn get_lowest_available_checkpoint(
-        &self,
-    ) -> sui_types::storage::error::Result<sui_types::messages_checkpoint::CheckpointSequenceNumber>
-    {
-        todo!()
-    }
-
-    fn get_checkpoint_by_digest(
-        &self,
-        _digest: &sui_types::messages_checkpoint::CheckpointDigest,
-    ) -> sui_types::storage::error::Result<Option<VerifiedCheckpoint>> {
-        todo!()
-    }
-
-    fn get_checkpoint_by_sequence_number(
-        &self,
-        sequence_number: sui_types::messages_checkpoint::CheckpointSequenceNumber,
-    ) -> sui_types::storage::error::Result<Option<VerifiedCheckpoint>> {
-        self.validator
-            .get_checkpoint_store()
-            .get_checkpoint_by_sequence_number(sequence_number)
-            .map_err(sui_types::storage::error::Error::custom)
-    }
-
-    fn get_checkpoint_contents_by_digest(
-        &self,
-        digest: &CheckpointContentsDigest,
-    ) -> sui_types::storage::error::Result<Option<sui_types::messages_checkpoint::CheckpointContents>>
-    {
-        self.validator
-            .get_checkpoint_store()
-            .get_checkpoint_contents(digest)
-            .map_err(sui_types::storage::error::Error::custom)
-    }
-
-    fn get_checkpoint_contents_by_sequence_number(
-        &self,
-        _sequence_number: sui_types::messages_checkpoint::CheckpointSequenceNumber,
-    ) -> sui_types::storage::error::Result<Option<sui_types::messages_checkpoint::CheckpointContents>>
-    {
-        todo!()
-    }
-
-    fn get_transaction(
-        &self,
-        tx_digest: &TransactionDigest,
-    ) -> sui_types::storage::error::Result<Option<Arc<sui_types::transaction::VerifiedTransaction>>>
-    {
-        self.validator
-            .get_cache_reader()
-            .get_transaction_block(tx_digest)
-            .map_err(sui_types::storage::error::Error::custom)
-    }
-
-    fn get_transaction_effects(
-        &self,
-        tx_digest: &TransactionDigest,
-    ) -> sui_types::storage::error::Result<Option<TransactionEffects>> {
-        self.validator
-            .get_cache_reader()
-            .get_executed_effects(tx_digest)
-            .map_err(sui_types::storage::error::Error::custom)
-    }
-
-    fn get_events(
-        &self,
-        event_digest: &TransactionEventsDigest,
-    ) -> sui_types::storage::error::Result<Option<TransactionEvents>> {
-        self.validator
-            .get_cache_reader()
-            .get_events(event_digest)
-            .map_err(sui_types::storage::error::Error::custom)
-    }
-
-    fn get_full_checkpoint_contents_by_sequence_number(
-        &self,
-        _sequence_number: sui_types::messages_checkpoint::CheckpointSequenceNumber,
-    ) -> sui_types::storage::error::Result<
-        Option<sui_types::messages_checkpoint::FullCheckpointContents>,
-    > {
-        todo!()
-    }
-
-    fn get_full_checkpoint_contents(
-        &self,
-        _digest: &CheckpointContentsDigest,
-    ) -> sui_types::storage::error::Result<
-        Option<sui_types::messages_checkpoint::FullCheckpointContents>,
-    > {
-        todo!()
-    }
-}
-
-impl ObjectStore for ValidatorWithFullnode {
-    fn get_object(
-        &self,
-        object_id: &ObjectID,
-    ) -> Result<Option<Object>, sui_types::storage::error::Error> {
-        self.validator.get_object_store().get_object(object_id)
-    }
-
-    fn get_object_by_key(
-        &self,
-        object_id: &ObjectID,
-        version: VersionNumber,
-    ) -> Result<Option<Object>, sui_types::storage::error::Error> {
-        self.validator
-            .get_object_store()
-            .get_object_by_key(object_id, version)
-    }
-}
-
-#[async_trait::async_trait]
-impl TransactionalAdapter for Simulacrum<StdRng, PersistedStore> {
+impl TransactionalAdapter for Simulacrum<StdRng, InMemoryStore> {
     async fn execute_txn(
         &mut self,
         transaction: Transaction,
